+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+### 4.0.0-alpha.0-1: 2025-12-22
+
+* Make dev mode authentic by using only `data-user-theme` attribute instead of artificial `.mastodon-bird-ui` class
+
+
 ### 4.0.0-alpha.0: 2025-12-21
 
 * Release install-to-mastodon.sh workflow